Drop utils file
